LayoutTests/platform/ios-sim-deprecated/fast/forms/input-select-webkit-user-select-none-expected.txt

Tests behavior of select() in case "-webkit-user-select: none" attribute is specified to the input element. The field should not be selected.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL input.select(); input.selectionStart should be 0. Was 10.
PASS input.selectionEnd is input.value.length
PASS input.value.length is not 0
PASS successfullyParsed is true

TEST COMPLETE
